scripts/consistency-rules/requirements.txt
annotated-types==0.6.0
blinker==1.6.3
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
typing_extensions==4.8.0
Werkzeug==3.0.0
pySigma==0.10.5
sigma-cli==0.7.7
pySigma-backend-splunk==1.0.3
rich==13.6.0
